Update LaTeX to handle optional parameters in one line per param context
The internal LaTeX code is based on a different logic than formerly:
now, it is the \sphinxparam which triggers a new line, not the commas.

Added some mark-up which will make possible customizing the style of the
commas.

Signatures with one line per parameter use currently a specific
\sphinxcode{,} which is customizable at user level.  Others use standard
commas, also customizable at user level.