Closes #11157: Keep "translated" attribute on translated nodes
`sphinx.transforms.Locale` iterates over all the nodes to replace its content
with the translated version of the text if found. Since the process works in two
phases, it adds an "internal" `translated: True` attribute to the docutils node
when the translated text is found in phase 1, allowing phase 2 to skip this
nodes and avoid re-processing them. Finally, this "internal" `translated`
attribute is removed from all the nodes.

This particular commit deletes the code that removes the `translated` attribute
from the nodes so developers can use it to improve the experience on
translations. This attribute can be useful to calculate the translated
percentage of a particular document, to visually mark paragraphs as not
translated and many other applications.