Change approach
The subclassing will not work because docutils doesn't use Sphinx's rst
parser for processing the included
rst text.  This is a shame, and
strictly speaking an issue in Sphinx/docutils.
So we go back to monkey patching.
It works, the change is now better contained inside Sphinx Include directive.
Added a unit test to verify that
includes of includes are correctly
handled.