coverage: Specify modules to run coverage for
Currently, the coverage builder lets you check for partially documented
modules, but there is no mechanism to identify totally undocumented
modules. Resolve this by introducing a new 'coverage_modules' config
option. This is a list of modules that should be documented somewhere
within the documentation tree. Any modules that are specified in the
configuration value but are not documented anywhere will result in a
warning. Likewise, any modules that are not in the config option but are
documented somewhere will result in a warning.

Signed-off-by: Stephen Finucane <stephen@that.guru>