sphinx.ext.coverage: Fix coverage for C APIs
We were using the 'objects' attribute of the C domain to pull all C
objects that we were aware of. However, this is never set to any value
nor does it ever appear to have been. Instead, we should have been using
and parsing the return value from 'get_objects'.

As an aside, this would suggest code coverage for C APIs has never
worked or at least has not worked for a *very* long time. This explains
why every GitHub hit for the various configuration options comes from
Sphinx (for the implementation), Python (for the sole user), or a fork
of these.

Signed-off-by: Stephen Finucane <stephen@that.guru>