Allow relative-to-root toctree dirs in autosummary
This treats a leading `/` in the directory of a `toctree` option in an
`autosummary` block as requesting a directory relative to the source
documentation root, rather than relative to the current directory.
Previously, prefixing a path with `/` would typically cause build
failures on Unix-likes because

    posixpath.join("/prefix/path", "/other") == "/other"

and Sphinx (hopefully!) didn't have permission to create directories in
the filesystem root.  This means that the change should not conflict
with any existing practical working documentation builds.

Autosummary currently uses `posixpath` for most operations even on
Windows, so the root indicator is only permitted as `/`, not `os.sep`.