Make test_include_source_read_event pass when run from /tmp
Because pytest's base tmp_path is also in /tmp, `source` here is
relative path, and thus is not properly converted to doc name.

So the test fails because `sources_reported` dict has a key like
`../../../pytest-of-debci/pytest-0/directive-include/baz/baz`, and
`assert "baz/baz" in sources_reported` fails.

For example, seen here:
https://salsa.debian.org/python-team/packages/sphinx/-/jobs/4612253