Leverage `importlib.reload` for reloading modules
Rather than directly manipulating `sys.modules` ourselves, rely on
`importlib.reload` to do it, as it works in more cases (particularly for
Rust extension modules built with PyO3).

Also, import modules with `typing.TYPE_CHECKING` set to `False` first,
and reload them with it set to `True`. This way, if the reload fails,
things are likely to still be in a valid state from the first,
successful import.