Do not use query components in URLs for assets in EPUB rendering
If the EPUB builder is detected, do not add checksums in a query
component to the URL of css and js assets.

The XHTML in EPUBs does not allow the use of query components in
relative URLs.
The checksumming of assets, introduced in
ae206694e68bea074aca633ea0d32e9ed882a95f unconditionally adds a query
component for the checksum, which breaks rendering on many enduser
devices (tested on e.g. Kindle Paperwhite 3, Kobo Aura H2O Edition 2).

Fixes #11598

Signed-off-by: David Runge <dave@sleepmap.de>