[C++] Resolve nested symbols through type aliases and base classes
Previously, symbol resolution always stopped at type aliases.  For
example, with the following document:

    .. cpp:class:: Class

       .. cpp::type:: NestedType = int

    .. cpp:type:: Alias = Class

    :cpp:expr:`Alias::NestedType`

`Alias::NestedType` was not able to be resolved.  With this change, it
is correctly resolved to `Class::NestedType`.

Additionally, symbols are also resolved through base classes in a
similar way:

    .. cpp:class:: Base

       .. cpp::type:: NestedType = int

    .. cpp:class:: Class : Base

    :cpp:expr:`Class::NestedType`

With this change, `Class::NestedType` will be correctly resolved to
`Base::NestedType`.