Fix singlehtml target uris to be same-document references
Before this change, target_uris would be generated as
`index.html#foo`.  This makes it difficult to rename the
generated document and can require reloading the document
when following a link, which can cause problems with JupyterLab
hosted documentation.

After this change, they are just `#foo`, which avoids the above
problems.