Fix loading custom template translations for en
The English language was hardcoded to load empty translations for
templates as part of https://github.com/sphinx-doc/sphinx/pull/10067,
but the rest of the code is already capable of handling missing
translations just fine, so we can simply drop the first if block.

This allows to use translations for English in custom templates, as for
all the other languages.

Fixes sphinx-doc/sphinx#12214