[linkcheck] expose hyperlink collection add-url call
The following tweaks the linkcheck builder's `HyperlinkCollector`
transform to process links with an `add_uri` method on the transform
(over an internal function inside the implementation). This can allow
extensions to derive and register additional transforms using
`HyperlinkCollector` to help inject hyperlinks into
`CheckExternalLinksBuilder` for custom node types that may have a
hyperlink value to check.

Signed-off-by: James Knight <james.d.knight@live.com>