[manpage] don't emit OSC 8 hyperlinks for anchor references (#12108)
A reference like ":ref:`Some other page <some-other-page>`" results
in a refuri "#some-other-page".  This does not seem useful to readers
of the man page. It is especially unhelpful when using a terminal
that implements a hint mode for selecting links -- the extra links
add noise, making it harder to select the interesting ones.
Don't emit OSC 8 for those.
Also don't emit it for URLs that might be unsafe
(see https://github.com/sphinx-doc/sphinx/pull/12260#issuecomment-2049147218)
I don't know one that would be unsafe but I'm sure there are some.

OTOH, "man_show_urls" doesn't seem unsafe because it shows the exact
URL that will be opened, so enable that for all URLs (except for
relative ones).

Follow up to #12108

I also confirmed that even with docutils 0.21 we do not need to
override depart_reference because we already skip reference nodes.