Fix #12380 (LaTeX footnotes)
The problem was there ever since #10191 added the footnote feature to
indicate the page number where footnote ends up printed, if not on same
page.

It is complicated to add a test for this change because the patch only
modifies how a certain custom LaTeX macro behaves so one has to check
PDF output which is bothersome.  But it got tested with a number of
sources already used for #10191 (one such test showed already the #12380
problem, but got overlooked at that time).

MEMO: the feature works only when Sphinx uses `\sphinxfootnotemark`
mark-up in LaTeX output; sometimes it may use rather a `footnote`
environment, then the mechanism is not activated.  This may happen for
example inside a cell of a long table, it can be that `Page N` will not
be added, because the `\sphinxfootnotemark` which does that is not
activated, although the footnote, which gets printed after the long
table ends, is on a different page.