[apidoc] Add `--remove-old` option
A common "gotcha" of re-running `sphinx-apidoc`, is that if the modules API changes it will not remove old files, leading to build errors.

Here we introduce a `-r/----remove-old` option,
to remove these files.

Note, a key detail here is that we don't want to simply clear the directory, before running `sphinx-apidoc`,
since this would lead to all files having a new `mtime`,
and so `sphinx-build` would rebuild all of them,
even if they have not changed.