✨ Add `sphinx.ext.apidoc` extension
A common usecase,
is that users simply want to point sphinx towards a Python module,
and have it generate documentation automatically.

This is not possible currently,
without a "pre-build" step of running the `sphinx-autogen` CLI.

This PR adds `sphinx.ext.apidoc` as a sphinx extension,
to incorporate the source file generation into the sphinx build.