ENH: Make sourcedir and outputdir parameters in sphinx-build optional
This allows to simply run `sphinx-build` in your doc directory,
instead of having to type `sphinx-build . _build`.

The default output folder name "_build" is a bit opinionated, but it's
what sphinx-quickstart uses. It creates that folder and adds it to
exclude_patterns in `conf.py` already. I believe, we can therefore
afford to go with this default. One may later add an `output_dir`
config variable to `conf.py` and use that instead of a hard-coded
default.