Slightly simplify toctree code
While working on #12720, I noticed that `parse_content` creates and
returns an empty node list. It does not do anything with it.

This PR moves the node out of `parse_content`, so that `parse_content`
only populates the toctree entries. The complete node creation logic
is now united in `TocTree.run`.