Glossary: derive index group from `category_key`
This fixes a bug where previously the `category_key` -- an optional
grouping that can be specified on glossary terms -- was used directly
as the group name to place terms within in the generated HTML index.

The fix re-uses a similar logic flow as used in `_key_func_1`, where
the `category_key` is used to derive (but not as an absolute source
of) the index group.