Glossary: nest `category_key` items in index
The grouping key (`category_key`) from glossary terms is used to
decide what symbol/character to group the corresponding terms into.

However: the terms are not guaranteed to begin with the same symbol
as the `category_key`.  This code therefore places them into a nested
definition entry in the index, with the complete `category_key` as
the parent label.

Relates to commit 4255bbe08ddb5d50db9780ad2e0d778d1f074d5a.