DOC: Highlight literals a bit more
The only style difference for literals is currently a monospace font.
This blends very much into surrounding text and makes it hard to read
the literals.

All other major themes (https://sphinx-themes.org/) use a background
for literals. It's a good idea to do the same in sphinx13. I've taken the
CSS from pydata-sphinx-theme, but I don't care too much about the actual
parameters as long as it stays visible but decent.
Note that I've limitied the background to `div.body` so that the background
is not applied in the sidebar. The sidebar entries are often single words
phrases and don't need a background. Actually, the backgound there would
be too cluttery.