Handle multiple inheritance correctly in autodoc
The previous code went through mro sequentially not treating it
as a tree, which is needed in case of multiple inheritance.
Take following example:
```python
class MyClass(Parent1, Parent2):
    foo_myclass=None

class Parent1(GrandParent1):
    foo_parent1=None

class GrantParent1:
    foo_grandparent1=None

class Parent2:
    foo_parent2
```

When `:show-inheritance: Parent1` is used, only following attributes should
be shown:
 * foo_myclass
 * foo_parent2

The foo_parent2 was previously not considered as the MRO matched Parent1 before
Parent2 but didn't check if it was really defined there.

The new code checks if the attribude is defined either directly in the ignored
class or in one of it's parents (that's why the issubclass in reversed order
is used).