LaTeX: fix a 7.4.0 typo in a default for \sphinxboxsetup
\sphinxbox and \sphinxboxsetup were added at 6.2.0. They are not used in
the LaTeX produced by Sphinx (nor are they tested by CI), they are only
provided for benefits of user redefinitions of default styling macros
for text, as a utility for decorative inline boxes.  At 7.4.0 the
default value for border-radius key which was \fboxsep got modified into
an explicit length which should have been 3pt but was written 0.4pt by
mistake.