Closes #12463: autosummary: fix empty __all__ is ignored
The check for the absence of `__all__` used the `all or dir(obj)`
syntax. Unfortunately, this resolves to the second value
not only for `None`, but also for an empty list. This lead to
the unintended behavior that setting `__all__ = []` was ignored
by autosummary. Checking explicitly for None fixes this.