Add `:include-subclasses:` option to `sphinx.ext.inheritance_diagram` (#8159)
When this option is given, all sub-classes of the classes listed as arguments to `inheritance-diagram` are included in the diagram.

This makes it possible to generate a complete inheritance tree by just listing the base class and adding `:include-subclasses:`.

This is different from specifying the module that contains this base class, as this module might include other classes that should not be part of the diagram.

Co-authored-by: Adam Turner <9087854+aa-turner@users.noreply.github.com>