Add the ``sphinx.ext.apidoc`` extension (#13333)
A common use-case is that users simply want to point Sphinx
towards a Python module, and have it generate documentation
automatically.

This is not possible currently, without a "pre-build" step
of running the ``sphinx-autogen`` CLI.

This PR adds ``sphinx.ext.apidoc`` as a first-party extension,
to incorporate the source file generation into the Sphinx build.

Co-authored-by: Adam Turner <9087854+aa-turner@users.noreply.github.com>