docs(extending_build.rst):  update to match example/todo.py
It appears `todo.py` was previously updated shifting some lines down
and this caused 5 code examples to be missing 1 or more lines at the end.
This was fixed, as well as replacing an old function with a new function
that parses directive content into nodes:  `parse_content_to_nodes()`.

Additionally, clarified when `doctree-resolved` event gets fired so make
the example more understandable.