LaTeX: allow more cases of table nesting, fix #13646
Tables using longtable can now contain nested tables inclusive of those
rendered by tabulary, up to the suppression of the latter horizontal
lines due to an upstream LaTeX bug.  A longtable can never itself be
nested, and will fall-back to tabular.

Formerly longtable would raise (in principle) an error if it contained
any sort of nested table, but the detection of being a longtable was
faulty if not specified as class option.

Relates #6838.