LaTeX: improve the table column widths computed by LaTeX itself
Fix #3447
Fix #8228

With this patch the legacy notion of "problematic cell" is almost
entirely removed, and LaTeX package tabulary will be used more often so
that column widths in tables may get better fit to actual contents.

This works well when the column contains cells with multiple paragraphs,
bullet lists, enumerated lists, or line blocks.  For cells with object
descriptions or admonitions, the column may not fit tightly to the
contents, but other columns will probably be improved.