Hyphenate versioning admonitions
Use e.g. 'version-changed' over 'versionchanged'. While the deprecated
admonition is marked as deprecated itself in the docs, we do not do
anything at runtime since there is almost zero cost to keeping this
relative to the work needed for millions of documents to be updated.

While here, we also update a use of the deprecated admonition with
version-changed since it makes more sense in the context (an attribute
of the feature was deprecated but not the whole feature itself).

Signed-off-by: Stephen Finucane <stephen@that.guru>