[singlehtml] toctree no filename with anchor
On #11970, same-document links dropped the uri (filename part).
On #12551, fixes to do not include the filename on hyperlinks was added.
So an anchor for a:
* doc reference: #document-my-page
* ref reference: #my-anchor.
On #13037, fix_refuris was removed, which would blindly rmatch #.* .
However, #12551 missed the toctree constructor, which was still
generating:
* #document-my-page#my-anchor.
Modify _resolve_toctree() to do not include rel_uri if the is an
anchorname.