LaTeX: fix #3099
Formerly, PDF rendering crashed for inputs of more than about 1350
codelines either from a code-block, or from a literal include, when this
literal include used the sphinxVerbatim and not the sphinxalltt route.

The fix goes via novel mark-up which is applied in case the literally
included file or code-block contains more than 500 codelines.  Then
chunks of 500 consecutive source codelines are rendered using:

    sphinxLongVerbatimFirst, sphinxLongVerbatimMiddle, and
    sphinxLongVerbatimLast.

Efforts have been devoted for line numbering and emphasizing to not be
modified, and (if so configured) for bottom captions to appear where
expected.

It was needed to (to cut into smaller pieces and) modify in a very
limited manner the legacy sphinxVerbatim which is still used for up to
500 codelines, which corresponds to about 10 a4-sized pages, if document
is with default pointsize and margins.

Hopefully no consequences will arise from these changes.

Known issue: in the unlikely event a pagebreak happens exactly after a
chunk of 500 code lines, the continuation hints usually printed at
bottom and top of pages are omitted.